+++ exif/log.go
@@ -1,5 +1,12 @@
 package exif
 
+import (
+	"fmt"
+
+	"github.com/evanoberholster/imagemeta/exif/ifds"
+	"github.com/evanoberholster/imagemeta/exif/tag"
+)
+
 // Inital Log Level
 var LogLevel = LogLevelError
 
@@ -12,7 +19,7 @@ const (
 )
 
 func checkLogLevel(l uint8) bool {
-	return LogLevel <= l
+	return LogLevel >= l
 }
 
 func isError() bool {
@@ -31,10 +38,14 @@ func isDebug() bool {
 	return checkLogLevel(LogLevelDebug)
 }
 
-func logTag() {
+func logTagInfo(ifd ifds.Ifd, t tag.Tag, offset uint32) {
+	fmt.Printf("Tag: %s\t Offset: x%.4x\t Name: %s\n", t, offset, ifd.TagName(t.ID))
+}
+
+func logIfdError() {
 
 }
 
-func logIfd() {
-
+func logIfdInfo(ifd ifds.Ifd, tagCount uint16, offset uint32) {
+	fmt.Println(ifd, tagCount, offset)
 }
