+++ exif/mknote.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/evanoberholster/imagemeta/exif/ifds"
 	"github.com/evanoberholster/imagemeta/exif/ifds/mknote"
+	"github.com/evanoberholster/imagemeta/imagetype"
 	"github.com/evanoberholster/imagemeta/meta"
 	"github.com/pkg/errors"
 )
@@ -55,4 +56,38 @@ func (r *reader) isNikonMkNoteHeader(ifd ifds.Ifd) (ifds.Ifd, binary.ByteOrder, 
 	}
 
 	return ifd, nil, ErrNikonMkNote
+}
+
+func (r *reader) parseMknoteIFD(e *Data, ifd ifds.Ifd) (ifds.Ifd, binary.ByteOrder) {
+	if e.make == "" {
+		return ifd, nil
+	}
+	//make := strings.ToUpper(e.make)
+	if e.make == "Canon" {
+		// Canon Makernotes do not have a Makernote Header
+		// offset 0
+		// ByteOrder is the same as RootIfd
+		return ifd, r.byteOrder
+	}
+	if e.make == "Nikon" || e.make == "NIKON CORPORATION" {
+		// Nikon v3 maker note is a self-contained Ifd
+		// (offsets are relative to the start of the maker note)
+
+		if ifd, byteOrder, err := r.isNikonMkNoteHeader(ifd); err == nil {
+			// update imagetype
+			if e.imageType == imagetype.ImageTiff {
+				e.imageType = imagetype.ImageNEF
+			}
+			return ifd, byteOrder
+		}
+		return ifd, nil
+	}
+	if e.make == "Sony" {
+		if e.imageType == imagetype.ImageTiff {
+			e.imageType = imagetype.ImageARW
+		}
+		return ifd, r.byteOrder
+	}
+
+	return ifd, nil
 }
